Add local events and listeners
Local events and listeners are purely support 
events that can be used outside of the event 
queue to split up game-logic for reusable
calculations (ie can these two players hear
each other when talking normally).

This required putting the listener_processor
in the local context, which is a bit 
inconsistent with how it's being used in some
places but nothing crazy.

Tested out the local events using the 
CommunicationEvent which is used in the 
communication module which is used in the
TalkEvent. This might need to be tweaked for
performance such that the server checks if
the talk event is valid then only sends it out
if appropriate