Implement starting the game
You can start the game with /start and see that you are assigned the 
test specialization with /specialization

/specialization has better usage information and now will also display 
the long description for commands if requested.

Locations now require a consecrated and lighting variable, both of which 
are currently unused.

Events also get passed the networking information in case they function 
by propagating more events. This is only necessary when you need to 
determine some information in a non-deterministic way (which requires 
one person decides and then tells everyone else the result)

World locations are setup in a very simplistic way.