Add simple bots & searching
Biggest change is that events triggered by an event now occur prior to 
other events. This occurred due to great difficulty adding bots due to 
race conditions, which the change resolved. This means that the wrapped 
event does literally nothing, so it has been removed

There is an issue where sometimes adventurers get the 'npc' 
specialization, but only rarely. There is currently a print statement to 
help catch what is happenning, but so far I've never had the print 
statement active and had it occur :/

For testing if things are occuring due to networking bugs, there is also 
a local_networking that can be used by starting with the local_game 
file.